NC51001 Sliding Window
链接：https://ac.nowcoder.com/acm/problem/51001
来源：牛客网

题目描述
An array of size n ≤ 106 is given to you. There is a sliding window of size k which is moving from the very left of the array to the very right. You can only see the k numbers in the window. Each time the sliding window moves rightwards by one position. Following is an example: 
The array is [1 3 -1 -3 5 3 6 7], and k is 3.
Window position 	Minimum value 	Maximum value
[1  3  -1] -3  5  3  6  7  	-1 	3
 1 [3  -1  -3] 5  3  6  7  	-3 	3
 1  3 [-1  -3  5] 3  6  7  	-3 	5
 1  3  -1 [-3  5  3] 6  7  	-3 	5
 1  3  -1  -3 [5  3  6] 7  	3 	6
 1  3  -1  -3  5 [3  6  7] 	3 	7
Your task is to determine the maximum and minimum values in the sliding window at each position. 

输入描述:

The input consists of two lines. The first line contains two integers n and k which are the lengths of the array and the sliding window. There are n integers in the second line.

输出描述:

There are two lines in the output. The first line gives the minimum values in the window at each position, from left to right, respectively. The second line gives the maximum values.

示例1
输入
复制

8 3
1 3 -1 -3 5 3 6 7

输出
复制

-1 -3 -3 -3 3 3
3 3 5 5 6 7